@@ -66,13 +66,17 @@
     "gulp-watch": "^4.3.8",
     "imports-loader": "^0.6.5",
     "jest": "^19.0.2",
+    "less": "^2.7.2",
+    "less-loader": "^4.0.4",
+    "node-sass": "^4.5.3",
     "npm-version": "^1.1.0",
     "react-lazy-render": "0.3.0",
     "react-test-renderer": "^15.5.4",
+    "sass-loader": "^6.0.6",
     "style-loader": "^0.13.1",
     "uglify-js": "2.4.x",
     "url-loader": "^0.5.7",
-    "webpack": "^2.0.0",
+    "webpack": "^3.0.0",
     "webpack-eslint-plugin": "^1.2.0"
   },
   "scripts": {
